requirements.txt
# Core framework
Django>=5.2,<5.3
djangorestframework>=3.16.1,<3.17

# Auth / social login
django-allauth[socialaccount]>=65.14.1,<66.0

# Runtime dependencies
requests>=2.32.0,<3.0
fontawesomefree>=6.6.0,<7.0
python-dotenv>=1.0.1,<2.0

# Deploy / production runtime
gunicorn>=22.0.0,<23.0
whitenoise>=6.7.0,<7.0
dj-database-url>=2.2.0,<3.0  # import: dj_database_url
django-redis>=5.4.0,<6.0
psycopg[binary]>=3.2.0,<4.0

# Legacy app referenced in INSTALLED_APPS
dr-scaffold>=2.1.2

# Build helpers (important for Python 3.14 environments)
setuptools>=75.0.0
wheel>=0.45.0
